PHShapePairForLCPにGetClosestPointを追加しました． それにともないCDShapeのclosestpointをとってくる必要があるので，CDShapeをUTRefでなくObject継承にしました． ContactDetectorは使ってない部分を消しました．
git-svn-id: http://springhead.info/spr2/Springhead2/trunk@3569 05cee5c3-a2e9-0310-9523-9dfc2f93dbe1

+++ src/Collision/CDCapsule.cpp
@@ -30,7 +30,7 @@ bool CDCapsule::IsInside(const Vec3f& p){
 		   (p.x*p.x + p.y*p.y < radius * radius && -0.5f*length < p.z && p.z < 0.5f*length);
 }
 float CDCapsule::CalcVolume(){
-	return  4.0f/3.0f * M_PI * radius * radius * radius + 
+	return  4.0f/3.0f * (float)M_PI * radius * radius * radius + 
 			(float)M_PI * radius * radius * length;
 }
